backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.98.0
azure-identity==1.15.0
pydantic[email]==2.11.7
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-multipart==0.0.6
sse-starlette==1.6.5
pgvector==0.2.4
langchain==0.3.27
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langchain-community
tiktoken==0.10.0
pydantic-settings
aiofiles==23.2.0
PyPDF==5.9.0
python-jose[cryptography]==3.3.0
bcrypt
azure-communication-email
